# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PICG API Documentation",
        "description": "This documentation covers all API endpoints for the PICG project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/education/delete": {
            "delete": {
                "tags": [
                    "Education"
                ],
                "summary": "Delete Education",
                "operationId": "deleteEducation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the education to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Education deleted successfully"
                    },
                    "404": {
                        "description": "Education not found"
                    },
                    "500": {
                        "description": "Failed to delete education"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/skill/delete": {
            "delete": {
                "tags": [
                    "Skill"
                ],
                "summary": "Delete a skill by ID",
                "operationId": "deleteSkill",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skill deleted successfully"
                    },
                    "404": {
                        "description": "Skill not found"
                    },
                    "500": {
                        "description": "Failed to delete skill"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}